// src/shared/interfaces/product-commets.ts
export interface ProductCommentInterface {
        id: number;
        content: string;
        productId: number;
        userId: string;
        createdAt: string;
        user: {
        id: number;
        name: string;
        email: string;
        avatar: {
            url: string;
        };
        rating: {
            value: number;
        };
    };
}